pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lsearch"
version = "0.1.0"
description = "Local RAG knowledge base for Claude Code - semantic + keyword hybrid search"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "moringchen", email = "moringchen@example.com"},
]
keywords = ["claude", "rag", "search", "knowledge-base", "mcp", "semantic-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "whoosh>=2.7.4",
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "markdown>=3.4.0",
    "tiktoken>=0.5.0",
    "networkx>=3.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
lsearch = "lsearch.cli:main"

[project.urls]
Homepage = "https://github.com/moringchen/lsearch"
Repository = "https://github.com/moringchen/lsearch"
Issues = "https://github.com/moringchen/lsearch/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lsearch"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
